Create Unittest for UpdateTaskCommandhandler MediatR
